==> 2. Deduplication.ipynb <==
import json
import pandas as pd
from pyspark.sql import SparkSession, Window
from pyspark.sql.functions import col,lower
import pyspark.sql.functions as F
from pyspark.ml.feature import RegexTokenizer, StopWordsRemover, Word2Vec

## 1. Read the cleaned dataset

with open('/Users/celine/Desktop/5430nlp/group/English_Only_More_than_10_reviews.json', 'r') as f:
    data = [json.loads(line) for line in f.readlines()]

reviews = [article['text'] for article in data]
len(reviews), len(set(reviews))

len(set(item['business_id'] for item in data))

## 2. Making copies of the original articles to avoid changing the imported data

feeds = []
for i, a in enumerate(data):
    feed = a.copy() 
    feed['review_id'] = i
    feeds.append(feed)

## 3. Run deduplication within each business

#### 1. Construct a list of list, each outer list is divided by business_id, each inter list is divided by review for that business_id

# Initialize a list to store reviews with the same business_id
business_id_list = []
BusinessReviews = []

# Iterate through the feeds to filter reviews with the same business_id
for feed in feeds:
    selBusinessID = feed['business_id']
    
    if selBusinessID not in business_id_list:
        
        sameBusinessReviews = [review for review in feeds if review['business_id'] == selBusinessID]
        BusinessReviews.append(sameBusinessReviews)
        business_id_list.append(selBusinessID)

len(BusinessReviews)

from simhash import Simhash, SimhashIndex

f = 64 # the number of bits, the default is 64

objs = {}  # a dictionary with business_id as keys and list of (review_id, simhash) as values

for business_reviews in BusinessReviews:
    business_id = str(business_reviews[0]['business_id'])  # Assuming 'business_id' is present in the first review
    objs[business_id] = []
    
    for review in business_reviews:
        review_id = str(review['review_id'])
        text = str(review['text'])
        sh = Simhash(text, f=f)
        objs[business_id].append((review_id, sh))
        
len(objs)

#123861
hamming_distance = 15

deduplicated_feed_data = []

for all_reviews in BusinessReviews:
    businessid = str(all_reviews[0]['business_id'])
    obj = objs.get(businessid, [])
    index = SimhashIndex(obj, k=hamming_distance, f=f)
    
    for business in all_reviews:
        selID = business['review_id']
        selFeed = feeds[selID]
        selHash = Simhash(str(selFeed['text']), f=64)
        
    # Get near duplicates using SimhashIndex
        duplicates = index.get_near_dups(selHash)
            
        if (len(duplicates) == 1) or (str(selID) == min(duplicates)):
            deduplicated_feed_data.append(selFeed)
                        
print(len(deduplicated_feed_data))

## 4. Evaluation of SimHash Performance

##### In summary, for text-based reviews, cosine similarity is often more widely used, especially when considering the overall content and importance of words in the reviews. 

# Convert deduplicated data to pandas dataframe
data = pd.DataFrame(deduplicated_feed_data)
# Creating a new DataFrame with selected columns
selected_columns = ['business_id', 'text']
df = data[selected_columns].copy()

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

# Function to calculate average cosine similarity for each business_id
def calculate_max_similarity(df):
    # Create an empty list to store the results
    result_data = []

    # Iterate over unique business_id values
    for business_id in df['business_id'].unique():
        # Filter DataFrame for the current business_id
        subset_df = df[df['business_id'] == business_id]

        # Use TF-IDF vectorization for the review texts
        tfidf_vectorizer = TfidfVectorizer()
        tfidf_matrix = tfidf_vectorizer.fit_transform(subset_df['text'])
        
        # Calculate cosine similarity matrix
        similarity_matrix = cosine_similarity(tfidf_matrix, tfidf_matrix)
        
        # Set diagonal elements to zero to exclude self-similarity
        np.fill_diagonal(similarity_matrix, 0)

        # Calculate the max similarity score for each business
        max_similarity = similarity_matrix.max()

        # Append the results to the list as a dictionary
        result_data.append({'business_id': business_id,
                            'max_similarity': max_similarity})

    # Create a DataFrame directly from the list of dictionaries
    result_df = pd.DataFrame(result_data)

    return result_df


# Example usage
result_df_max_similarity = calculate_max_similarity(df)
result_df_max_similarity = result_df_max_similarity.sort_values(by='max_similarity', ascending=False)
result_df_max_similarity.head()

# Check the business with highest similarity score and its review, ensures no duplicated reviews
pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None)
df222 = df[df['business_id'] == '_Q3XmBesceFqn9Dk-w6tlA'].sort_values(by='text', ascending=False)
df222.head()

## 5. Export Json File

with open('/Users/celine/Desktop/5430nlp/group/deduplicated_reviews.json', 'w') as f:
    for article in deduplicated_feed_data:
        json.dump(article, f)
        f.write('\n')
print(f'Number of non-duplicate articles: {len(deduplicated_feed_data)}')
