==> 3. Text_Summarization.ipynb <==
## 1. Read the dataset

#### a. Import data

import pandas as pd
data = pd.read_json('/Users/celine/Desktop/5430nlp/group/deduplicated_reviews.json', lines = True)
print(data.shape[0])

pd.set_option('display.max_columns', None)
data.head()

#### b. Select necessary columns only

# Selecting specific columns
selected_columns = ['business_id', 'address', 'name','postal_code', 'stars', 'review_id', 'text','useful','customer_review_stars']

# Creating a new DataFrame with selected columns
df = data[selected_columns].copy()

#### c. Tokenize text column and remove stopwords

from nltk.tokenize import RegexpTokenizer
from nltk.corpus import stopwords

tokenizer = RegexpTokenizer(r'\w+')
stop_words = set(stopwords.words('english'))
df['tokenize_text'] = df['text'].apply(lambda x: [token for token in tokenizer.tokenize(x.lower()) if token.lower() not in stop_words])
df.head()

#### d. Add a column with the number of tokenized words

df['text_length'] = df['tokenize_text'].apply(len)
df.head()

## 2. Split reviews into positive, neutral and negative

positive_review = df[df['customer_review_stars'] > 3]
neutral_review = df[df['customer_review_stars'] == 3]
negative_review = df[df['customer_review_stars'] < 3]

print(positive_review.shape[0])
print(neutral_review.shape[0])
print(negative_review.shape[0])

positive_review.head()

## 3. For each business, select the top 20 most representative text

#### a. Sort the dataframe based on useful and length of text

##### Reasons: useful represents the number of likes by other users that find this review useful. However, many business does not have 20 voted for useful. Thus, we want to select the reviews with more words as useful.
##### Assumption: The longer the review, the more representative it is.

positive_review = positive_review.sort_values(by=['business_id', 'useful', 'text_length'], ascending=[False, False, False])
neutral_review = neutral_review.sort_values(by=['business_id', 'useful', 'text_length'], ascending=[False, False, False])
negative_review = negative_review.sort_values(by=['business_id', 'useful', 'text_length'], ascending=[False, False, False])

positive_review.head()

#### b. Find the top 20 reviews for each business' subset reviews

###### 1. Positive

# Rank reviews for each business based on the value of useful
positive_review['useful_rank'] = positive_review.groupby('business_id')['useful'].rank(ascending=False, method='min')
positive_review['positive_review_count'] = positive_review.groupby('business_id')['review_id'].transform('count')
positive_review['useful_positive_review_count'] = positive_review.groupby('business_id')['useful'].transform(lambda x: (x > 0).sum())
print(positive_review.shape)

positive_review = positive_review[positive_review['useful_rank'] <= 20]
print(positive_review.shape)

# Group by 'business_id' and aggregate other columns using 'first()'
grouped_positive_df = positive_review.groupby('business_id').agg({
    'address': 'first',
    'name': 'first',
    'postal_code': 'first',
    'stars': 'first',
    'text': lambda x: ' '.join(x),  # Aggregating the 'text' column as before
    'positive_review_count': 'first',
    'useful_positive_review_count': 'first'
}).reset_index()

grouped_positive_df['text_length'] = grouped_positive_df['text'].apply(len)
grouped_positive_df = grouped_positive_df.sort_values('text_length', ascending = False)
print(grouped_positive_df.shape)
grouped_positive_df.head(30)


###### 2. Neutral

# Rank reviews for each business based on the value of useful
neutral_review['useful_rank'] = neutral_review.groupby('business_id')['useful'].rank(ascending=False, method='min')
neutral_review['neutral_review_count'] = neutral_review.groupby('business_id')['review_id'].transform('count')
neutral_review['useful_neutral_review_count'] = neutral_review.groupby('business_id')['useful'].transform(lambda x: (x > 0).sum())
print(neutral_review.shape)

neutral_review = neutral_review[neutral_review['useful_rank'] <= 20]
print(neutral_review.shape)

# Group by 'business_id' and aggregate other columns using 'first()'
grouped_neutral_df = neutral_review.groupby('business_id').agg({
    'address': 'first',
    'name': 'first',
    'postal_code': 'first',
    'stars': 'first',
    'text': lambda x: ' '.join(x),  # Aggregating the 'text' column as before
    'neutral_review_count': 'first',
    'useful_neutral_review_count': 'first'
}).reset_index()

grouped_neutral_df['text_length'] = grouped_neutral_df['text'].apply(len)
grouped_neutral_df = grouped_neutral_df.sort_values('text_length', ascending = False)
print(grouped_neutral_df.shape)
grouped_neutral_df.head(30)


###### 3. Negative

# Rank reviews for each business based on the value of useful
negative_review['useful_rank'] = negative_review.groupby('business_id')['useful'].rank(ascending=False, method='min')
negative_review['negative_review_count'] = negative_review.groupby('business_id')['review_id'].transform('count')
negative_review['useful_negative_review_count'] = negative_review.groupby('business_id')['useful'].transform(lambda x: (x > 0).sum())
print(negative_review.shape)

negative_review = negative_review[negative_review['useful_rank'] <= 20]
print(negative_review.shape)

# Group by 'business_id' and aggregate other columns using 'first()'
grouped_negative_df = negative_review.groupby('business_id').agg({
    'address': 'first',
    'name': 'first',
    'postal_code': 'first',
    'stars': 'first',
    'text': lambda x: ' '.join(x),  # Aggregating the 'text' column as before
    'negative_review_count': 'first',
    'useful_negative_review_count': 'first'
}).reset_index()

grouped_negative_df['text_length'] = grouped_negative_df['text'].apply(len)
grouped_negative_df = grouped_negative_df.sort_values('text_length', ascending = False)
print(grouped_negative_df.shape)
grouped_negative_df.head(30)


## 4. Group reviews by business_id and perform summarization

from sumy.parsers.plaintext import PlaintextParser
from sumy.nlp.tokenizers import Tokenizer
from sumy.summarizers.lex_rank import LexRankSummarizer
from sumy.summarizers.text_rank import TextRankSummarizer
from sumy.summarizers.lsa import LsaSummarizer
from sumy.summarizers.kl import KLSummarizer
from sumy.summarizers.luhn import LuhnSummarizer

def LexRank_summary(text):
    NUM_SENTS = 2
    summary = str()

    parser = PlaintextParser.from_string(text, Tokenizer("english"))
    summarizer = LexRankSummarizer()

    sentences = summarizer(parser.document, NUM_SENTS)  # Summarize the document with 2 sentences
    for sentence in sentences:
        summary += (sentence.__unicode__()+ ' ')


    return summary

# Apply the generate_summary function to each row in the DataFrame
grouped_positive_df['positive_summarization'] = grouped_positive_df['text'].apply(LexRank_summary)
grouped_neutral_df['neutral_summarization'] = grouped_neutral_df['text'].apply(LexRank_summary)
grouped_negative_df['negative_summarization'] = grouped_negative_df['text'].apply(LexRank_summary)

# Display the DataFrame with the new "summarization" column
grouped_positive_df.head()


## 5. Combine Positive, Neutral, and Negative for each business

merged_df = pd.merge(grouped_positive_df[['business_id', 'positive_review_count', 'useful_positive_review_count', 'positive_summarization']],
                     grouped_neutral_df[['business_id','neutral_review_count','useful_neutral_review_count','neutral_summarization']],
                     on='business_id',
                     how='outer')

merged_df = pd.merge(merged_df, 
                     grouped_negative_df[['business_id','negative_review_count','useful_negative_review_count','negative_summarization']], 
                     on='business_id', 
                     how='outer')
merged_df.head()

merged_df['positive_summarization'].fillna('Sorry, this business has no positive reviews', inplace=True)
merged_df['neutral_summarization'].fillna('Sorry, this business has no neutral reviews', inplace=True)
merged_df['negative_summarization'].fillna('Sorry, this business has no negative reviews', inplace=True)
merged_df.fillna(0, inplace=True)

## 6. Evaluation using Sentiment Analysis

import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer 

sent_analyzer = SentimentIntensityAnalyzer()

eva_df = merged_df.copy()

eva_df['pos_analyzer'] = eva_df['positive_summarization'].apply(lambda x: sent_analyzer.polarity_scores(x)['compound'])
eva_df['pos_analyzer'].mean()

eva_df['neu_analyzer'] = eva_df['neutral_summarization'].apply(lambda x: sent_analyzer.polarity_scores(x)['compound'])
eva_df['neu_analyzer'].mean()

eva_df['neg_analyzer'] = eva_df['negative_summarization'].apply(lambda x: sent_analyzer.polarity_scores(x)['compound'])
eva_df['neg_analyzer'].mean()

eva_df.head()
