==> 4. Word_Embedding_Model.ipynb <==
from pyspark.sql import SparkSession
import pyspark.sql.functions as F
from pyspark.sql.window import Window

### Initialize PySpark

spark = SparkSession.builder.appName('wordembed').getOrCreate()

### Load data into Spark dataframe

wembed_df = spark.read.json('/Users/celine/Desktop/5430nlp/group/deduplicated_reviews.json')
wembed_df.count()

wembed_df = wembed_df.select('business_id','name','text')
wembed_df.show(5)

wembed_df = wembed_df.withColumn('inputText', F.col('text'))
wembed_df.show(5)

### Train a Word2Vec Model

from pyspark.ml.feature import RegexTokenizer, StopWordsRemover, Word2Vec

# regular expression tokenizer to tokenize inputText into individual tokens (words)
regextok = RegexTokenizer(gaps = False, pattern = '\w+',
                          inputCol = 'inputText', outputCol = 'tokens')
# StopWordsRemover to remove stopwords in the list of tokens
stopwrmv = StopWordsRemover(inputCol = 'tokens', outputCol = 'tokens_sw_removed')

wembed_df = regextok.transform(wembed_df)
wembed_df = stopwrmv.transform(wembed_df)
wembed_df.show(5)

word2vec = Word2Vec(vectorSize = 100, minCount = 5, inputCol = 'tokens_sw_removed', outputCol = 'wordvectors')
wembed_model = word2vec.fit(wembed_df)


model_path = "/Users/celine/Desktop/5430nlp/group/web/word_embedding_model"
wembed_model.save(model_path)


# apply model on data wembed_df
wembed_df2 = wembed_model.transform(wembed_df)
wembed_df2.show(5)

wembed_df2.select('business_id','name','text','wordvectors').write.format("parquet").save("/Users/celine/Desktop/5430nlp/group/web/dataset_with_vectors.parquet")
