==> 5. Topic_Modelling.ipynb <==
import pandas as pd
import json
import nltk
from gensim.models import CoherenceModel
from nltk.tokenize import word_tokenize
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import GridSearchCV

with open('deduplicated_reviews 2.json', 'r') as f:
    data =  [json.loads(line.strip()) for line in f.readlines()]

# Data cleaning

columns_to_keep = ['business_id', 'name', 'stars', 'address', 'city', 'postal_code', 'text']
df = df[columns_to_keep]

#combine all the reviews together
combined_text = df.groupby('business_id')['text'].apply(' '.join).reset_index()

# Merging the aggregated 'text' with the original data based on 'name'
df = pd.merge(df.drop('text', axis=1).drop_duplicates(subset=['business_id']), combined_text, on='business_id')

# Printing the resulting DataFrame
print(df)

# Step 1: Convert 'text' column to lowercase
df['text'] = df['text'].str.lower()

# Step 2: Remove specified words from 'text' column
words_to_remove = ['bite', 'bit', 'restaurant', 'business', 'dish', 'meal', 'wait', 'waiter', 'waitress', 'favorite',
                   'person', 'manager', 'size', 'point', 'husband', 'dine', 'course', 'home', 'ice', 'item', 'line',
                   'location', 'owner', 'portion', 'selection', 'server', 'service', 'food', 'street', 'table',
                   'taste', 'friend', 'plate', 'bowl', 'water', 'week', 'weekend', 'space', 'room', 'friend']
for word in words_to_remove:
    df['text'] = df['text'].str.replace(r'\b' + word + r'\b', '', regex=True)

# Step 3: Tokenize and lemmatize, keeping only nouns
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')

lemmatizer = WordNetLemmatizer()
def get_wordnet_pos(word):
    tag = nltk.pos_tag([word])[0][1][0].upper()
    tag_dict = {"J": wordnet.ADJ,
                "N": wordnet.NOUN,
                "V": wordnet.VERB,
                "R": wordnet.ADV}
    return tag_dict.get(tag, wordnet.NOUN)

def lemmatize_text(text):
    tokens = word_tokenize(text)
    lemmatized_tokens = [lemmatizer.lemmatize(token, get_wordnet_pos(token)) for token in tokens]
    nouns = [token for token, pos in nltk.pos_tag(lemmatized_tokens) if pos.startswith('N')]
    return ' '.join(nouns)

df['text'] = df['text'].apply(lemmatize_text)

# Step 4: Create clusters
def get_cluster(text):
    if 'japanese' in text:
        return 'japanese'
    elif 'pizza' in text:
        return 'pizza'
    elif 'burger' in text:
        return 'burger'
    elif 'korean' in text:
        return 'korean'
    elif 'italian' in text:
        return 'italian'
    elif 'american' in text:
        return 'american'
    elif 'chinese' in text:
        return 'chinese'
    elif 'vegan' in text:
        return 'vegan'
    elif 'thai' in text:
        return 'thai'
    else:
        return 'others'

df['cluster'] = df['text'].apply(get_cluster)

print(df)

cluster_counts = df['cluster'].value_counts()
print(cluster_counts)

df.to_csv('processed_data.csv')

# Topic modelling

param_grid = {
    'n_components': [5, 10, 15],  # Try different numbers of topics
    'learning_method': ['batch', 'online'],
    'max_iter': [100, 500, 1000] 
    # Add other parameters to the grid
}

# Create CountVectorizer instance
tf_vectorizer = CountVectorizer(strip_accents='unicode', 
                                stop_words='english',
                                lowercase=True,
                                max_features=100,
                                token_pattern=r'\b[a-zA-Z]{3,}\b',
                                max_df=0.7,
                                min_df=100,
                                ngram_range=(1, 3))

# Transform the text data
tf = tf_vectorizer.fit_transform(df['text'])

# Create LDA model
lda = LatentDirichletAllocation()

# GridSearchCV for parameter tuning
grid_search = GridSearchCV(lda, param_grid=param_grid, cv=5, n_jobs=-1, verbose=2)
grid_search.fit(tf)  

# Get the results as a DataFrame
results = pd.DataFrame(grid_search.cv_results_)

# Calculate perplexity for each parameter combination
perplexities = []
for params in grid_search.cv_results_['params']:
    lda_model = LatentDirichletAllocation(**params)
    lda_model.fit(tf)
    perplexity = lda_model.perplexity(tf)
    perplexities.append(perplexity)

# Add perplexity scores to the results DataFrame
results['perplexity'] = perplexities

# Display the table with perplexity scores for each parameter combination
print(results[['param_n_components', 'param_learning_method', 'param_max_iter', 'perplexity']])

# Best parameters
print(grid_search.best_params_)

#best LDA model
lda_tf = LatentDirichletAllocation(n_components=15, 
                                   max_iter=1000,
                                   learning_method='online', 
                                   random_state = 28474460)
lda_tf.fit(tf)

#Examine the LDA features
tf_feature_names = tf_vectorizer.get_feature_names_out()

print('Number of features is:', len(tf_feature_names))
print('First 10 features:', tf_feature_names[:10])

import pyLDAvis
import pyLDAvis.lda_model
pyLDAvis.enable_notebook()
pyLDAvis.lda_model.prepare(lda_tf, tf, tf_vectorizer)

n_top_words = 3  # Number of top words to extract

# Extract top words for each review
reviews = {}
for idx, review in enumerate(tf):  # Assuming 'tf' is your topic-token matrix
    topic = lda_tf.transform(review.reshape(1, -1))  # Transform the review to match LDA input
    ordered_indexes = topic.argsort()[0][::-1]
    top_indexes = ordered_indexes[:n_top_words]
    top_features = [tf_feature_names[i] for i in top_indexes]
    reviews[idx] = top_features

# Add a new 'topic' column to df1 containing the top 3 words for each review
df['topic'] = [reviews[i] for i in range(len(df))]

print(df)

# Convert the first character of each word in the 'topic' column to uppercase
df['topic'] = df['topic'].apply(lambda x: [word.capitalize() for word in x])
print(df)

df.to_json('tm.json', orient='records')
