==> 6. Topic_Summarization_Photo.ipynb <==
from pyspark.sql import SparkSession
import pyspark.sql.functions as F
from pyspark.sql.window import Window
from pyspark.sql.functions import lit, concat_ws


## Load data from photo, topic modeling, text summarization

spark= SparkSession.builder.getOrCreate()

# Load data from photo.json, topic modeling dataset, text summarization dataset
photo = spark.read.json('/Users/celine/Desktop/5430nlp/group/photos.json')
topic = spark.read.json('/Users/celine/Desktop/5430nlp/group/tm.json')
summarization = spark.read.option("header", "true") \
    .option("multiLine", "true") \
    .option("quote", "\"") \
    .option("escape", "\"").csv('/Users/celine/Desktop/5430nlp/group/summarization_data_clean.csv')


photo.show()

topic.show()

summarization.show()

## Photo

# Select photo with label food and drink
photo_food = photo.filter((F.col('label') == 'food') | (F.col('label') == 'drink'))


# Only select one photo for each business
windowSpec = Window.partitionBy('business_id').orderBy('photo_id')

# Add a row number column
photo_food_with_row_number = photo_food.withColumn('row_number', F.row_number().over(windowSpec))

# Filter for the first photo for each business
photo_per_business = photo_food_with_row_number.filter(F.col('row_number') == 1)

# Now you can drop the 'row_number' column if you want
photo_per_business = photo_per_business.drop('row_number')

photo_per_business.count()


## Topic modeling

# Add a state column fill "PA"
topic = topic.withColumn('state', lit('PA'))
# Concat address, city, state, postal_code as full address
topic = topic.withColumn('full_address', concat_ws(', ', 'address', 'city', concat_ws(' ', 'state','postal_code')))

topic.show()

topic = topic.select('business_id', 'name', 'stars', 'full_address', 'topic')

# Join photo and topic modeling dataset
photo_food_tm = photo_per_business.select('business_id', 'photo_id').join(topic, 'business_id', 'right')
photo_food_tm.count()

photo_food_tm.show()

# Text summarization

# Join photo_food_tm and text summarization dataset
topic_sum_photo = photo_food_tm.join(summarization, 'business_id', 'inner')
topic_sum_photo.show()
